improved username validation
